requirements-dev.txt
-r requirements.txt
pytest==5.4.1
pytest-flask==1.0.0
pytest-mock==3.1.0
pytest-cov==2.8.1
